requirements.txt
# Core Dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.3.0,<5.0.0
django-filter>=23.3,<24.0
django-redis>=5.4.0,<6.0.0
redis>=5.0.0,<6.0.0
celery>=5.3.0,<6.0.0
gunicorn>=21.2.0,<22.0.0
psycopg2-binary>=2.9.9,<3.0.0
Pillow>=10.0.0,<11.0.0
python-dotenv>=1.0.0,<2.0.0
drf-yasg>=1.21.7,<2.0.0
drf-spectacular==0.27.0
django-storages>=1.14.2,<2.0.0
boto3>=1.34.34,<2.0.0
botocore>=1.34.162,<2.0.0
asgiref>=3.8.1,<4.0.0
paystackapi>=2.1.0,<3.0.0  # Paystack API integration
requests>=2.31.0,<3.0.0
whitenoise>=6.6.0,<7.0.0

# WebSocket Support for Chat
channels>=4.0.0,<5.0.0
channels-redis>=4.1.0,<5.0.0

# Development Dependencies
django-debug-toolbar>=4.2.0,<5.0.0
pytest>=7.4.0,<8.0.0
pytest-django>=4.7.0,<5.0.0
coverage>=7.3.0,<8.0.0
black>=23.9.0,<24.0.0
flake8>=6.1.0,<7.0.0
isort>=5.12.0,<6.0.0

# Security
django-axes>=6.3.0,<7.0.0
django-ratelimit>=4.1.0,<5.0.0

# API Documentation
drf-spectacular>=0.26.0,<1.0.0

# Task Queue
celery[redis]>=5.3.0,<6.0.0
flower>=2.0.0,<3.0.0  # Celery monitoring
django-celery-beat>=2.5.0,<3.0.0  # Database-backed periodic tasks

# Caching
django-redis>=5.4.0,<6.0.0

# Email
django-anymail>=10.1,<11.0

# Monitoring
sentry-sdk>=1.32.0,<2.0.0

# Testing
factory-boy>=3.3.0,<4.0.0
Faker>=22.6.0

# Additional Dependencies
django-phonenumber-field
qrcode[pil]
python-barcode[images]
django-auth-ldap>=4.0.0,<5.0.0
pycryptodome>=3.0
cryptography>=41.0.0

# File processing and security
python-magic>=0.4.27
Pillow>=10.0.0
cryptography>=41.0.0

